+++ tests/e2e/utils/src/factories.js
@@ -1,6 +1,9 @@
 import { HTTPClientFactory } from '@woocommerce/api';
 const config = require( 'config' );
 import { simpleProductFactory } from './factories/simple-product';
+import { variableProductFactory } from './factories/variable-product';
+import { variationFactory } from './factories/variation';
+import { groupedProductFactory } from './factories/grouped-product';
 
 const apiUrl = config.get( 'url' );
 const adminUsername = config.get( 'users.admin.username' );
@@ -20,6 +23,9 @@ const factories = {
 	},
 	products: {
 		simple: simpleProductFactory( withDefaultPermalinks ),
+		variable: variableProductFactory( withDefaultPermalinks ),
+		variation: variationFactory( withDefaultPermalinks ),
+		grouped: groupedProductFactory( withDefaultPermalinks )
 	},
 };