+++ tests/legacy/bootstrap.php
@@ -58,6 +58,9 @@ class WC_Unit_Tests_Bootstrap {
 		// load test function so tests_add_filter() is available.
 		require_once $this->wp_tests_dir . '/includes/functions.php';
 
+		// Always load PayPal Standard for unit tests.
+		tests_add_filter( 'woocommerce_should_load_paypal_standard', '__return_true' );
+
 		// load WC.
 		tests_add_filter( 'muplugins_loaded', array( $this, 'load_wc' ) );
 
@@ -174,11 +177,11 @@ class WC_Unit_Tests_Bootstrap {
 		define( 'WC_REMOVE_ALL_DATA', true );
 		include $this->plugin_dir . '/uninstall.php';
 
-		WC_Install::install();
-
 		// Initialize the WC API extensions.
 		\Automattic\WooCommerce\Admin\Install::create_tables();
 		\Automattic\WooCommerce\Admin\Install::create_events();
+
+		WC_Install::install();
 
 		// Reload capabilities after install, see https://core.trac.wordpress.org/ticket/28374.
 		if ( version_compare( $GLOBALS['wp_version'], '4.7', '<' ) ) {